Export `core::entity_serde` with custom serde functions for entity
Looks nicer when placed in a single module. Also I need them for the upcoming networking triggers.
